nit: use Option<i32> instead of Option<u8> for compression level (#308)

+++ p2p/src/authenticated/mod.rs
@@ -149,7 +149,7 @@
 //!
 //!     // Register some channel
 //!     const MAX_MESSAGE_BACKLOG: usize = 128;
-//!     const COMPRESSION_LEVEL: Option<u8> = Some(3);
+//!     const COMPRESSION_LEVEL: Option<i32> = Some(3);
 //!     let (sender, receiver) = network.register(
 //!         0,
 //!         Quota::per_second(NonZeroU32::new(1).unwrap()),
@@ -548,7 +548,7 @@ mod tests {
         });
     }
 
-    fn test_message_too_large(compression: Option<u8>) {
+    fn test_message_too_large(compression: Option<i32>) {
         // Configure test
         let base_port = 3000;
         let n: usize = 2;
